+++ packages/mobile/src/api/article.ts
@@ -14,3 +14,15 @@ export const usePopularArticle = () => {
   );
   return response;
 };
+
+export const fetchArticleByBoard = (boardId: number) => {
+  return caxios.get<res.ArticleByBoard[]>(`/boards/${boardId}/articles`);
+};
+
+export const useArticleByBoard = (boardId: number) => {
+  const response = useQuery<AxiosResponse<res.ArticleByBoard[]>, Error>(
+    [ "articles", boardId ],
+    () => fetchArticleByBoard(boardId),
+  );
+  return response;
+};
