--- packages/mobile/src/type/apiTypes/board.d.ts
@@ -1,10 +1,0 @@
-declare namespace res {
-  type ArticleByBoard = {
-    id: number;
-    board: Board;
-    title: string;
-    hits: number;
-    scraps: number;
-    dates: string;
-  };
-}
